Se definió e implementó la función respaldar() en la clase laboratorio. En el arhivo main.cpp, se hizo uso de este método para almacenar el contenido del objeto lab en un archivo .txt llamado: laboratorio.txt

+++ computadora.cpp
@@ -32,7 +32,7 @@ void Computadora::setRAM(float ram)
 {
     this->ram = ram;
 }
-int Computadora::getRAM()
+float Computadora::getRAM()
 {
     return ram;
 }
